+++ odoo/local-src/sf_rma/hooks.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
+from odoo import api, SUPERUSER_ID
+
+
+def post_init_hook(cr, registry):
+    env = api.Environment(cr, SUPERUSER_ID, {})
+
+    warehouses = env['stock.warehouse'].search([])
+    warehouses.create_locations_rma()