+++ odoo/local-src/sf_rma/models/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import models
+from . import mrp_repair
+from . import picking
+from . import rma
+from . import rma_cause
+from . import sale
+from . import stock_move
+from . import warehouse