+++ HISTORY.rst
@@ -24,6 +24,9 @@ latest (unreleased)
 * Add functional field on supplier invoice showing duplicated invoices
 * Smart button on supplier invoice showing duplicated invoices
 * BSSFL-65: Add LDAP configuration
+* Add RMA module
+  Create new object sf.rma to link with mrp.repair, sale.order and stock.picking.  
+  This object will be used from zendesk.
 
 **Bugfixes**